requirements.txt
discord.py
loguru
python-dotenv